feat: Add sorting algorithms
This commit adds several sorting algorithms to the project, including insertion sort, bubble sort, recursive insertion sort, recursive bubble sort, selection sort, and quick sort. Each algorithm is implemented in a separate file and can be used to sort an array of integers. The time and space complexities for each algorithm are also provided in the comments.